2020.12.17.

+++ routes/article.js
@@ -14,21 +14,33 @@ router.route('/:id')
     });
 })
 .delete(async(req, res) => {
-    try {
-        await articleModel.deleteOne({ _id: req.params.id });
-        res.render('index', { 
+    if (req.body.password == process.env.DELETE_KEY)
+    {
+        try {
+            await articleModel.deleteOne({ _id: req.params.id });
+            res.render('index', { 
+                layout: path.join('layouts', 'with_jumbotron'),
+                title: 'Home',
+                nav: '1',
+                articles: await articleModel.find().lean(),
+                articleDelete: true
+              });
+        } 
+        catch (err) {
+            res.render('error', { 
+                title: 'Error',
+                nav: '1',
+                specificMessage: err.message
+            });
+        }
+    }
+    else {
+        res.render('article', { 
             layout: path.join('layouts', 'with_jumbotron'),
-            title: 'Home',
+            title: 'Article',
             nav: '1',
-            articles: await articleModel.find().lean(),
-            articleDelete: true
-          });
-    } 
-    catch (err) {
-        res.render('error', { 
-            title: 'Error',
-            nav: '1',
-            specificMessage: err.message
+            article: await articleModel.findById(req.params.id),
+            wrongSecret: true
         });
     }
 });
